+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -110,6 +110,7 @@ object HybrisConstants {
     const val EXTENSIONS_XML = "extensions.xml"
     const val COCKPIT_NG_DEFINITION_XML = "definition.xml"
     const val HYBRIS_LICENCE_JAR = "hybrislicence.jar"
+    const val SAP_LICENCES = "installedSaplicenses.properties"
 
     const val HYBRIS_DIRECTORY = "hybris"
     const val HYBRIS_DATA_DIRECTORY = "data"